requirement.txt
# Framework
pytest
pytest-html
selenium
Appium-Python-Client
pycmd

# Reporting
allure-pytest

# Data Driven
pyexcel
pyexcel-xls
pyexcel-xlsx

# Image Draw/Signature Draw
Pillow
image